+++ nym-connect/src/components/ConnectionStatus.tsx
@@ -3,32 +3,65 @@ import { Box, CircularProgress, Tooltip, Typography } from '@mui/material';
 import { DateTime } from 'luxon';
 import { ConnectionStatusKind, GatewayPerformance } from '../types';
 import { ServiceProvider } from '../types/directory';
-import { ServiceProviderInfo } from './ServiceProviderInfo';
+import { GatwayWarningInfo, ServiceProviderInfo } from './TooltipInfo';
+import { ErrorOutline, InfoOutlined } from '@mui/icons-material';
 
-const FONT_SIZE = '10px';
+const FONT_SIZE = '14px';
 const FONT_WEIGHT = '600';
 const FONT_STYLE = 'normal';
 
 const ConnectionStatusContent: FCWithChildren<{
   status: ConnectionStatusKind;
-}> = ({ status }) => {
+  serviceProvider?: ServiceProvider;
+  gatewayError: boolean;
+}> = ({ status, serviceProvider, gatewayError }) => {
+  if (gatewayError) {
+    return (
+      <Tooltip title={serviceProvider ? <GatwayWarningInfo /> : undefined}>
+        <Box
+          display="flex"
+          alignItems="center"
+          gap={0.5}
+          justifyContent="center"
+          sx={{ cursor: 'pointer' }}
+          color="warning.main"
+        >
+          <ErrorOutline sx={{ fontSize: 14 }} />
+          <Typography fontWeight={FONT_WEIGHT} fontStyle={FONT_STYLE} fontSize={FONT_SIZE} textAlign="center">
+            Gateway has issues
+          </Typography>
+        </Box>
+      </Tooltip>
+    );
+  }
   switch (status) {
-    case ConnectionStatusKind.connected:
+    case 'connected':
       return (
-        <Typography fontWeight={FONT_WEIGHT} fontStyle={FONT_STYLE} fontSize="14px">
-          Connected to
+        <Tooltip title={serviceProvider ? <ServiceProviderInfo serviceProvider={serviceProvider} /> : undefined}>
+          <Box display="flex" alignItems="center" gap={0.5} justifyContent="center" sx={{ cursor: 'pointer' }}>
+            <InfoOutlined sx={{ fontSize: 14 }} />
+            <Typography fontWeight={FONT_WEIGHT} fontStyle={FONT_STYLE} fontSize={FONT_SIZE} textAlign="center">
+              Connected to Nym Mixnet
+            </Typography>
+          </Box>
+        </Tooltip>
+      );
+    case 'disconnected':
+      return (
+        <Typography fontWeight={FONT_WEIGHT} fontStyle={FONT_STYLE} textAlign="center" fontSize={FONT_SIZE}>
+          Connect to the mixnet
         </Typography>
       );
-    case ConnectionStatusKind.disconnecting:
+    case 'disconnecting':
       return (
         <Box display="flex" alignItems="center" justifyContent="center">
           <CircularProgress size={FONT_SIZE} color="inherit" />
-          <Typography fontWeight={FONT_WEIGHT} fontStyle={FONT_STYLE} ml={1}>
+          <Typography fontWeight={FONT_WEIGHT} fontStyle={FONT_STYLE}>
             Disconnecting...
           </Typography>
         </Box>
       );
-    case ConnectionStatusKind.connecting:
+    case 'connecting':
       return (
         <Box display="flex" alignItems="center" justifyContent="center">
           <CircularProgress size={FONT_SIZE} color="inherit" />
@@ -37,20 +70,7 @@ const ConnectionStatusContent: FCWithChildren<{
           </Typography>
         </Box>
       );
-    case ConnectionStatusKind.disconnected:
-      return (
-        <Typography
-          fontWeight={FONT_WEIGHT}
-          fontStyle={FONT_STYLE}
-          ml={1}
-          textTransform="uppercase"
-          textAlign="center"
-          fontSize={FONT_SIZE}
-          sx={{ wordSpacing: 3, letterSpacing: 2 }}
-        >
-          You are not protected
-        </Typography>
-      );
+
     default:
       return null;
   }
@@ -62,29 +82,18 @@ export const ConnectionStatus: FCWithChildren<{
   connectedSince?: DateTime;
   serviceProvider?: ServiceProvider;
 }> = ({ status, serviceProvider, gatewayPerformance }) => {
-  const color =
-    status === ConnectionStatusKind.connected || status === ConnectionStatusKind.disconnecting
-      ? '#21D072'
-      : 'warning.main';
+  const color = status === 'connected' || status === 'disconnecting' ? '#21D072' : 'white';
+  console.log(gatewayPerformance);
 
   return (
     <>
-      <Box color={color} fontSize={FONT_SIZE} sx={{ mb: 1 }}>
-        {status === ConnectionStatusKind.connected && gatewayPerformance !== 'Good' ? (
-          <Typography fontWeight={FONT_WEIGHT} fontStyle={FONT_STYLE} textAlign="left" color="primary">
-            Gateway has issues
-          </Typography>
-        ) : (
-          <ConnectionStatusContent status={status} />
-        )}
+      <Box color={color} sx={{ mb: 2 }}>
+        <ConnectionStatusContent
+          status={status}
+          serviceProvider={serviceProvider}
+          gatewayError={gatewayPerformance !== 'Good'}
+        />
       </Box>
-      {serviceProvider ? (
-        <Tooltip title={<ServiceProviderInfo serviceProvider={serviceProvider} />}>
-          <Box sx={{ cursor: 'pointer' }}>
-            {serviceProvider && <Typography>{serviceProvider.description}</Typography>}
-          </Box>
-        </Tooltip>
-      ) : null}
     </>
   );
 };