+++ nym-connect/src/components/TooltipInfo.tsx
@@ -7,14 +7,23 @@ export const ServiceProviderInfo = ({ serviceProvider }: { serviceProvider: Serv
     <Typography variant="body2" fontWeight="bold">
       Connection info
     </Typography>
-    <Typography variant="caption">{serviceProvider.description}</Typography>
     <Divider />
     <Typography variant="caption" fontWeight="bold">
       Gateway <Typography variant="caption">{serviceProvider.gateway}</Typography>
     </Typography>
     <Divider />
     <Typography variant="caption" fontWeight="bold">
-      Provider <Typography variant="caption">{serviceProvider.address.slice(0, 35)}...</Typography>
+      Service provider <Typography variant="caption">{serviceProvider.address.slice(0, 35)}...</Typography>
     </Typography>
+  </Stack>
+);
+
+export const GatwayWarningInfo = () => (
+  <Stack gap={1} sx={{ wordWrap: 'break-word', maxWidth: 150, p: 1 }}>
+    <Typography variant="body2" fontWeight="bold" color="warning.main">
+      Connection issue
+    </Typography>
+    <Divider />
+    <Typography variant="caption">Try disconnecting and connecting again</Typography>
   </Stack>
 );